+++ pkg/analysis_server/lib/src/services/refactoring/legacy/extract_widget.dart
@@ -636,7 +636,7 @@ class _ParametersCollector extends RecursiveAstVisitor<void> {
         );
       }
     } else if (element is LocalVariableElement) {
-      if (!expressionRange.contains(element.firstFragment.nameOffset)) {
+      if (!expressionRange.contains(element.firstFragment.nameOffset2!)) {
         if (node.inSetterContext()) {
           status.addError("Write to '$elementName' cannot be extracted.");
         } else {