+++ pkg/analyzer/CHANGELOG.md
@@ -8,6 +8,7 @@
 * Remove deprecated `PromotableElement` and `PromotableFragment`.
 * Remove deprecated `LocalVariableElement.hasInitializer`.
 * Remove deprecated `LocalVariableFragment.hasInitializer`.
+* Remove deprecated `LocalVariableFragment.nameOffset`.
 * `LibraryFragment` is not `Annotatable` anymore.
 * Stop implementing `ConstantEvaluationTarget` by `ElementAnnotation`.
   This is an internal interface that should not have been exposed.